backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
crewai==0.28.8
langchain
langchain-openai
langchain-google-genai
firebase-admin==6.4.0
python-multipart==0.0.6
pydantic==2.5.0
python-dotenv==1.0.0
sendgrid==6.10.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
pydantic-settings==2.1.0
